refactor: folders, unused code
