+++ module/JepRiaShowcase/App/client-react/src/features/feature/state/FeatureReducer.ts
@@ -1,11 +1,8 @@
-import { Feature, FeatureSearchTemplate } from "./../features/feature/api/FeatureTypes";
+import { Feature, FeatureSearchTemplate } from "../api/FeatureTypes";
 import { EntityState, OptionState, SearchState } from "@jfront/core-redux-thunk";
-import { combineReducers, Reducer } from "@reduxjs/toolkit";
-import { initialEntityState, reducer as crudReducer } from "../features/feature/state/featureSlice";
-import {
-  initialSearchState,
-  reducer as searchReducer,
-} from "../features/feature/featureSearchSlice";
+import { combineReducers } from "@reduxjs/toolkit";
+import { initialEntityState, reducer as crudReducer } from "./FeatureSlice";
+import { initialSearchState, reducer as searchReducer } from "./FeatureSearchSlice";
 
 export interface FeatureState {
   feature: {
